enterprise-logistics/services/shipment-service/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
psycopg2-binary
python-jose
python-dotenv
pydantic[email]
python-multipart
pytest
httpx
pytest-cov
